refactor(tui): extract sticky prompt prototype
